UI、演出周りの実装

+++ Assets/Scripts/CharacterInfo.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 [CreateAssetMenu(menuName = "ScriptableObject/CharacterInfo")]
 public class CharacterInfo:ScriptableObject{
     public CharacterType characterType;
-    public float recoverHP;
+    public int recoverHP;
     public ToolsController toolsController;
 }
 public enum CharacterType{
